backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyJWT==2.8.0

# ML dependencies
numpy==2.0.2
pandas==2.2.0
scipy==1.14.0
scikit-learn==1.6.1
xgboost==2.0.0
mlflow==3.1.4
cloudpickle==3.1.1
boto3==1.29.7  # For S3 artifact storage
joblib==1.3.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
hiredis==2.2.3  # C parser for Redis

# Data validation
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0

# Monitoring and metrics
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
isort==5.12.0

# Additional utilities
python-json-logger==2.0.7
tenacity==8.2.3  # Retry logic
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2023.3